Refactor: add modularization in Statements Printer
