+++ TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.cs
@@ -5,6 +5,8 @@ using ApprovalTests;
 using ApprovalTests.Reporters;
 using TheatricalPlayersRefactoringKata.Core;
 using TheatricalPlayersRefactoringKata.Core.Entities;
+using TheatricalPlayersRefactoringKata.Core.Interfaces;
+using TheatricalPlayersRefactoringKata.Core.Printers;
 using Xunit;
 
 #endregion
@@ -17,7 +19,8 @@ public class StatementPrinterTests
     [UseReporter(typeof(DiffReporter))]
     public void TestStatementExampleLegacy()
     {
-        var plays = new Dictionary<string, Play> {
+        var plays = new Dictionary<string, Play>
+        {
             { "hamlet", new Play("Hamlet", 4024, "tragedy") },
             { "as-like", new Play("As You Like It", 2670, "comedy") },
             { "othello", new Play("Othello", 3560, "tragedy") }
@@ -32,7 +35,7 @@ public class StatementPrinterTests
             ]
         );
 
-        var result = StatementPrinter.Print(invoice);
+        var result = TextStatementPrinter.Print(invoice);
 
         Approvals.Verify(result);
     }
@@ -41,7 +44,8 @@ public class StatementPrinterTests
     [UseReporter(typeof(DiffReporter))]
     public void TestTextStatementExample()
     {
-        var plays = new Dictionary<string, Play> {
+        var plays = new Dictionary<string, Play>
+        {
             { "hamlet", new Play("Hamlet", 4024, "tragedy") },
             { "as-like", new Play("As You Like It", 2670, "comedy") },
             { "othello", new Play("Othello", 3560, "tragedy") },
@@ -62,7 +66,7 @@ public class StatementPrinterTests
             ]
         );
 
-        var result = StatementPrinter.Print(invoice);
+        var result = TextStatementPrinter.Print(invoice);
 
         Approvals.Verify(result);
     }
@@ -71,7 +75,8 @@ public class StatementPrinterTests
     [UseReporter(typeof(DiffReporter))]
     public void TestXmlStatementExample()
     {
-        var plays = new Dictionary<string, Play> {
+        var plays = new Dictionary<string, Play>
+        {
             { "hamlet", new Play("Hamlet", 4024, "tragedy") },
             { "as-like", new Play("As You Like It", 2670, "comedy") },
             { "othello", new Play("Othello", 3560, "tragedy") },
@@ -92,7 +97,7 @@ public class StatementPrinterTests
             ]
         );
 
-        var result = StatementPrinter.XmlPrint(invoice);
+        var result = XmlStatementPrinter.Print(invoice);
 
         Approvals.Verify(result);
     }
