+++ TheatricalPlayersRefactoringKata/Core/Interfaces/IPlay.cs
@@ -5,6 +5,6 @@ public interface IPlay
     public string Name { get; }
 
     public string Type { get; }
-    
+
     public int CalculateAmount(int audience);
 }
