+++ TheatricalPlayersRefactoringKata/Core/Interfaces/IStatementPrinter.cs
@@ -1,0 +1,8 @@
+﻿using TheatricalPlayersRefactoringKata.Core.Entities;
+
+namespace TheatricalPlayersRefactoringKata.Core.Interfaces;
+
+public interface IStatementPrinter
+{
+    public static abstract string Print(Invoice invoice);
+}
